+++ src/provider/Diagnostic/setBaseDiag.ts
@@ -22,7 +22,7 @@ function baseDiagnostic(DocStrMap: TTokenStream, AST: TAstRoot): readonly CDiagB
 
     const diagList: readonly CDiagBase[] = [
         ...getLineErr(DocStrMap),
-        ...getTreeErr(AST, displayErrList, DocStrMap),
+        ...getTreeErr(AST, displayErrList),
         ...getMultilineDiag(DocStrMap),
     ];
     wm.set(DocStrMap, diagList);