+++ src/provider/Diagnostic/tools/TreeErr/getDirectivesErr.ts
@@ -1,9 +1,9 @@
+/* eslint-disable max-lines */
 /* eslint-disable max-lines-per-function */
 import * as vscode from 'vscode';
 import { CAhkDirectives } from '../../../../AhkSymbol/CAhkLine';
 import type { TAhkSymbol } from '../../../../AhkSymbol/TAhkSymbolIn';
 import { EDiagCode } from '../../../../diag';
-import type { TTokenStream } from '../../../../globalEnum';
 import type { THashTagUPKey } from '../../../../tools/Built-in/Directives';
 import { DirectivesMDMap } from '../../../../tools/Built-in/Directives';
 import { CDiagBase } from '../CDiagBase';
@@ -80,39 +80,7 @@ const DiagDirectivesMap: ReadonlyMap<string, TDiagMsg> = ((): ReadonlyMap<string
     return map;
 })();
 
-/**
- * ```ahk
- * #Hotstring EndChars
- * ;            ^ No support
- * ```
- *
- * [read more](https://www.autohotkey.com/docs/commands/_Hotstring.htm)
- */
-function HotstringEndChars(ch: CAhkDirectives, DocStrMap: TTokenStream): CDiagBase[] {
-    const { selectionRange } = ch;
-    const { lStr } = DocStrMap[selectionRange.start.line];
-
-    if (!(/^\s*#Hotstring\b[ \t]+\bEndChars\b[ \t]/ui).test(lStr)) {
-        return [];
-    }
-    // OK --> #Hotstring NoMouse
-    // NG --> #Hotstring EndChars NewChars
-    // OK --> #Hotstring NewOptions
-
-    // did i need flag support of NewOptions
-    // https://www.autohotkey.com/docs/Hotstrings.htm#Options
-
-    return [
-        new CDiagBase({
-            value: EDiagCode.code904,
-            range: selectionRange,
-            severity: vscode.DiagnosticSeverity.Warning,
-            tags: [],
-        }),
-    ];
-}
-
-export function getDirectivesErr(ch: TAhkSymbol, DocStrMap: TTokenStream): CDiagBase[] {
+export function getDirectivesErr(ch: TAhkSymbol): CDiagBase[] {
     // err of #Directives
     if (!(ch instanceof CAhkDirectives)) return [];
 
@@ -128,10 +96,6 @@ export function getDirectivesErr(ch: TAhkSymbol, DocStrMap: TTokenStream): CDiag
                 tags,
             }),
         ];
-    }
-
-    if (hashtag === 'HOTSTRING') {
-        return HotstringEndChars(ch, DocStrMap);
     }
 
     // check is unknown Directives or not